Moved add term to history to its own file
